move shale.py to shale folder
